Unnecessary quotation marks in text of GUI Version
The material in the echo task does not require quotes; the guiversion variable set in the call to ant already had quotes.

There was also still a mistake in the straight Linux build, Richard: you use ./version.sh to generate the guiversion string when doing make GUI. But there is no such ./version.sh unless it gets copied or linked to from CommonBuildResources.